+++ src/database/operations/fiscalSponsorships/index.ts
@@ -1,0 +1,2 @@
+export * from './createOrUpdateFiscalSponsorship';
+export * from './removeFiscalSponsorship';