+++ src/database/operations/index.ts
@@ -8,6 +8,7 @@ export * from './changemakerProposals';
 export * from './changemakers';
 export * from './dataProviders';
 export * from './ephemeralUserGroupAssociations';
+export * from './fiscalSponsorships';
 export * from './funders';
 export * from './generic';
 export * from './opportunities';